LongSales

+++ src/main/java/ru/netology/stats/Java6/services/Main.java
@@ -6,7 +6,7 @@ public class Main {
     public static void main (String args []){
         StatsService service = new StatsService();
 
-              int[] monthlySalesArray = {8, 15, 13, 15, 17, 20, 19, 20, 7, 14, 14, 18};
+              long[] monthlySalesArray = {8, 15, 13, 15, 17, 20, 19, 20, 7, 14, 14, 18};
