+++ src/test/java/ru/netology/stats/Java6/services/StatsServiceTest.java
@@ -8,7 +8,7 @@ public class StatsServiceTest {
     @Test
     public void TestAllSales() {
         StatsService service = new StatsService();
-        int[] monthlySalesArray = {8, 15, 13, 15, 17, 20, 19, 20, 7, 14, 14, 18};
+        long[] monthlySalesArray = {8, 15, 13, 15, 17, 20, 19, 20, 7, 14, 14, 18};
 
         int expected = 180;
         long actual = service.allSales(monthlySalesArray);
@@ -20,7 +20,7 @@ public class StatsServiceTest {
     @Test
     public void TestAverageSales() {
         StatsService service = new StatsService();
-        int[] monthlySalesArray = {8, 15, 13, 15, 17, 20, 19, 20, 7, 14, 14, 18};
+        long[] monthlySalesArray = {8, 15, 13, 15, 17, 20, 19, 20, 7, 14, 14, 18};
 
         int expected = 15;
         long actual = service.averageSales(monthlySalesArray);
@@ -32,10 +32,10 @@ public class StatsServiceTest {
     @Test
     public void TestMinSales() {
         StatsService service = new StatsService();
-        int[] monthlySalesArray = {8, 15, 13, 15, 17, 20, 19, 20, 7, 14, 14, 18};
+        long[] monthlySalesArray = {8, 15, 13, 15, 17, 20, 19, 20, 7, 14, 14, 18};
 
         int expected = 9;
-        int actual = service.minSales(monthlySalesArray);
+        long actual = service.minSales(monthlySalesArray);
 
         Assertions.assertEquals(expected,actual);
     }
@@ -43,10 +43,10 @@ public class StatsServiceTest {
     @Test
     public void TestMaxSales() {
         StatsService service = new StatsService();
-        int[] monthlySalesArray = {8, 15, 13, 15, 17, 20, 19, 20, 7, 14, 14, 18};
+        long[] monthlySalesArray = {8, 15, 13, 15, 17, 20, 19, 20, 7, 14, 14, 18};
 
         int expected = 8;
-        int actual = service.maxSales(monthlySalesArray);
+        long actual = service.maxSales(monthlySalesArray);
 
         Assertions.assertEquals(expected,actual);
     }
@@ -54,7 +54,7 @@ public class StatsServiceTest {
     @Test
     public void TestMonthHighSales() {
         StatsService service = new StatsService();
-        int[] monthlySalesArray = {8, 15, 13, 15, 17, 20, 19, 20, 7, 14, 14, 18};
+        long[] monthlySalesArray = {8, 15, 13, 15, 17, 20, 19, 20, 7, 14, 14, 18};
 
         int expected = 5;
         long actual = service.monthHighSales(monthlySalesArray);
@@ -65,7 +65,7 @@ public class StatsServiceTest {
     @Test
     public void TestMonthLowSales() {
         StatsService service = new StatsService();
-        int[] monthlySalesArray = {8, 15, 13, 15, 17, 20, 19, 20, 7, 14, 14, 18};
+        long[] monthlySalesArray = {8, 15, 13, 15, 17, 20, 19, 20, 7, 14, 14, 18};
 
         int expected = 5;
         long actual = service.monthLowSales(monthlySalesArray);
